requirements.txt
appdirs==1.4.4
asgiref==3.7.2
bcrypt==4.1.2
build==1.0.3
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cryptography==42.0.5
cycler==0.11.0
Cython==0.29.19
decorator==4.4.2
distlib==0.3.8
Django==5.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
docutils==0.19
duckdb==0.7.1
Faker==22.0.0
filelock==3.13.1
fonttools==4.39.3
future==0.18.3
gurux-common==1.0.13
gurux-dlms==1.0.136
gurux-net==1.0.19
gurux-serial==1.0.20
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
iso8601==1.1.0
Jinja2==3.1.3
joblib==1.2.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.7.0
kivy-deps.sdl2-dev==0.6.0
Kivy-examples==2.0.0rc3
Kivy-Garden==0.1.5
kivy-garden.xcamera==2020.613
kivy-garden.zbarcam==2020.524
kivymd==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.5
matplotlib==3.7.1
moviepy==1.0.3
mysqlclient==2.2.1
numpy==1.24.2
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.0
pandas==2.0.1
paramiko==3.4.0
Pillow==9.5.0
platformdirs==4.1.0
plyer==2.1.0
proglog==0.1.10
pyarrow==11.0.0
pycparser==2.21
Pygments==2.14.0
pyjnius==1.6.1
pylance==0.4.4
PyNaCl==1.5.0
pyparsing==3.0.9
pypiwin32==223
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-docx==0.8.11
python-for-android==2024.1.21
pytz==2023.3
pywin32==305
PyYAML==6.0
pyzbar==0.1.9
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
serial==0.0.97
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.66.2
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.25.0
xcamera==2020.613
XlsxWriter==3.0.3
zbarcam==2020.524
